+++ src/useDidMount.spec.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, cleanup as cleanupReact } from '@testing-library/react';
 import { cleanup as cleanupHooks, renderHook } from '@testing-library/react-hooks';
-import useOnMount from './useOnMount';
+import useDidMount from './useDidMount';
 
-describe('useOnMount', () => {
+describe('useDidMount', () => {
   beforeEach(() => {
     cleanupHooks();
     cleanupReact();
@@ -11,21 +11,21 @@ describe('useOnMount', () => {
   });
 
   it('should be an arrow function', () => {
-    expect(useOnMount).to.be.a('function');
-    expect(useOnMount.prototype).to.be.empty;
+    expect(useDidMount).to.be.a('function');
+    expect(useDidMount.prototype).to.be.empty;
   });
 
   it('should return a single function', () => {
-    const { result } = renderHook(() => useOnMount());
+    const { result } = renderHook(() => useDidMount());
 
     expect(result.current).to.be.a('function');
   });
 
-  it('the returned function should be a setter for a callback to be performed when component mounts', () => {
+  it('the returned function should be a setter for a callback to be performed when component did mount', () => {
     const spy = sinon.spy();
 
     const TestComponent = () => {
-      const onMount = useOnMount();
+      const onMount = useDidMount();
 
       onMount(spy);